+++ bld/src/old_win/datactl/combo.c
@@ -49,7 +49,7 @@ BOOL ctl_combo_finish( ctl_elt *elt, HANDLE inst, HWND dlg, void *ptr )
     inst = inst;
 
     _value_int( ptr, elt ) = elt->info.combo.origin +
-            SendDlgItemMessage( dlg, elt->control,
+            (int)SendDlgItemMessage( dlg, elt->control,
                         ctl_combo_get_msg( dlg, elt->control ), NULL, NULL );
 
     return( TRUE );
